Added test dependencies to pom.xml, Modified States, User, UserRepository. Added new classes Team, Role

+++ src/main/java/se/azza/userservice/services/UserService.java
@@ -4,10 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
+import se.azza.userservice.model.Issue;
 import se.azza.userservice.model.User;
 import se.azza.userservice.repository.UserRepository;
+import se.azza.userservice.resttemplates.RestTemplates;
 
 @Service
 public class UserService {
@@ -19,16 +24,26 @@ public class UserService {
 		return userRepository.save(user);
 	}
 
-	public Optional<User> getUserById(long userId) {
-		return userRepository.findById(userId);
+	public ResponseEntity<User> getUserById(long userId) {
+		Optional<User> user = userRepository.findById(userId);
+		if(!user.isEmpty()) {
+			return new ResponseEntity<User>(user.get(), HttpStatus.OK);			
+		}
+		return new ResponseEntity<>(HttpStatus.NOT_FOUND); 
 	}
 
 	public void updateUser(User user) {
 		userRepository.save(user);
 	}
 
-	public void deleteById(Long id) {
-		userRepository.deleteById(id);
+	public ResponseEntity<String> deleteById(Long id, RestTemplate restTemplate) {
+		List<Issue> issuesForUser = RestTemplates.getIssuesForUser(restTemplate, id);
+		if (issuesForUser.isEmpty()) {
+			userRepository.deleteById(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+		return ResponseEntity.badRequest()
+				.body("You need to delete all issues first... then you available to delete user");
 	}
 
 	public List<User> getAllUsers() {
